improve steps fragment (now it is not download again when user rotates device, tab is always correct showed in a lot of number of steps lessons)
