+++ app/src/main/java/org/stepic/droid/base/FragmentBase.java
@@ -19,6 +19,7 @@ import org.stepic.droid.store.IDownloadManager;
 import org.stepic.droid.store.operations.DatabaseFacade;
 import org.stepic.droid.util.resolvers.CoursePropertyResolver;
 import org.stepic.droid.util.resolvers.ISearchResolver;
+import org.stepic.droid.util.resolvers.IStepResolver;
 import org.stepic.droid.util.resolvers.IVideoResolver;
 
 import javax.inject.Inject;
@@ -64,6 +65,9 @@ public class FragmentBase extends Fragment {
 
     @Inject
     public CoursePropertyResolver mCoursePropertyResolver;
+
+    @Inject
+    public IStepResolver mStepResolver;
 
 
     public FragmentBase() {
